+++ src/main/java/presenters/recipeBuilder/RecipeBuilderPresenter.java
@@ -1,4 +1,4 @@
-package presenters;
+package presenters.recipeBuilder;
 
 import useCases.RecipeBuilderResponseModel;
